refactoring - separate Syntax classes out
