+++ catbird/cbird.py
@@ -1,11 +1,9 @@
 from abc import ABC
 from collections.abc import Iterable
 from copy import deepcopy
-import json
 import numpy as np
-from pathlib import Path
-import subprocess
 from .syntax import SyntaxPath
+from .utils import json_from_exec
 
 type_mapping = {'Integer' : int,
                 'Boolean' : bool,
@@ -328,86 +326,15 @@ class Catbird(ABC):
             print(attr_str)
 
 
-def json_from_exec(exec):
+def problems_from_json(json_obj, problem_names=None):
     """
     Returns the Python objects corresponding to the MOOSE application described
     by the json file.
 
     Parameters
     ----------
-    json_file : str, or Path
-        Either an open file handle, or a path to the json file. If `json` is a
-        dict, it is assumed this is a pre-parsed json object.
-
-    Returns
-    -------
-    dict
-        A dictionary of all MOOSE objects
-    """
-    json_proc = subprocess.Popen([exec, '--json'], stdout=subprocess.PIPE)
-    json_str = ''
-
-    # filter out the header and footer from the json data
-    while True:
-        line = json_proc.stdout.readline().decode()
-        if not line:
-            break
-        if '**START JSON DATA**' in line:
-            continue
-        if '**END JSON DATA**' in line:
-            continue
-
-        json_str += line
-
-    j_obj = json.loads(json_str)
-
-    return j_obj
-
-def write_json(json_dict_out,name):
-    """
-    Write a dictionary in JSON format
-
-    Parameters
-    ----------
-    json_dict_out : dict
-    name: str
-      Save as name.json
-    """
-    json_output = json.dumps(json_dict_out, indent=4)
-    json_name=name
-    if json_name.find(".json") < 0 :
-        json_name = name+".json"
-
-    with open(json_name, "w") as fh:
-        fh.write(json_output)
-        fh.write("\n")
-    print("Wrote to ",json_name)
-
-
-def read_json(json_file):
-    """
-    Load the contents of a JSON file into a dict.
-
-    Parameters
-    ----------
-    json_file: str
-      Name of JSON file
-    """
-    json_dict = {}
-    with open(json_file) as handle:
-        json_dict = json.load(handle)
-    return json_dict
-
-def problems_from_json(json_file, problem_names=None):
-    """
-    Returns the Python objects corresponding to the MOOSE application described
-    by the json file.
-
-    Parameters
-    ----------
-    json_file : dict, str, or Path
-        Either an open file handle, or a path to the json file. If `json` is a
-        dict, it is assumed this is a pre-parsed json object.
+    json_obj : dict
+        Pre-parsed json object containing MOOSE syntax
     problems : Iterable of str
         Set of problems to generate classes for
 
@@ -417,15 +344,10 @@ def problems_from_json(json_file, problem_names=None):
         A dictionary of problem objects
     """
 
-    if isinstance(json_file, dict):
-        json_obj = json_file
-    else:
-        json_obj = json.load(json_file)
+    assert isinstance(json_obj, dict)
 
     out = dict()
-
     out['problems'] = parse_problems(json_obj, problem_names=problem_names)
-
     return out
 
 def parse_blocks(json_obj):
