+++ catbird/factory.py
@@ -1,5 +1,6 @@
 from .syntax import SyntaxRegistry
-from .cbird import parse_block, read_json, write_json, json_from_exec
+from .cbird import parse_block
+from .utils import read_json, write_json, json_from_exec
 
 class Factory():
     def __init__(self,exec_path,config_file=None):
